[process] Hold reference to process during call to step()
It is conceivable that the process may terminate during the execution of
step().  If nothing else holds a reference to the containing object, this
would cause the object to be freed prior to returning from step().

Add a ref_get()/ref_put() around the call to ->step() to prevent this
from happening.