[tests] Increase timeout for QEMU commands
Loading a snapshot can take a significant amount of time if it is not
already cached, since it can be several hundred megabytes.  Allow 15
seconds instead of 5 to handle this case.

This number should probably be tweaked in the future, or an additional
"wait forever" option for loading VMs should be added.