[bofm] Calculate mport value manually
BOFM firmware doesn't actually populate the mport field with a
meaningful value.  In order to support cards with multiple ports per
PCI bus:dev.fn address, we must manually scan through the BOFM EN
entries and count how many times our PCI bus:dev.fn address has
appeared in order to determine the effective mport value.

This is not foolproof, since it requires the EN entries to be in
ascending port order.  An alternative method would involve using the
{slot,port} number to impose the correct ordering on the EN entries,
but this would still not be foolproof since it wouldn't handle the
case of an EN entry existing for only the second of two ports.  (Also,
most BOFM versions will zero out the {slot,port} number at the time
that the PCI bus:dev.fn address is filled in, so the information may
not be available anyway.)

Signed-off-by: Michael Brown <mcb30@ipxe.org>