[efi] Wait for link state to be established during Initialize()
Some broken software (including the reference implementation on
tianocore.org) assumes that the link state is valid immediately after
calling Initialize().  This neither conforms to the specification for
Initialize(), nor reflects the reality of any networking hardware more
recent than an RS-232 serial port.

Work around this broken software by delaying until the link state can
be established.  This unfortunately makes Initialize() (and Reset())
into calls that can potentially block for a noticeable length of time.
This is hideously ugly (not least because there is no way for a user
to abort this delay).

A proper solution would be to fix the code on tianocore.org to allow
for the fact that link detection is not instantaneous.

Signed-off-by: Michael Brown <mcb30@ipxe.org>