[intel] Always use autoloaded MAC address if it is valid
Some Intel NICs (e.g. the flashless variants of the i210) use
interestingly obscure ways to store the MAC address in order to reduce
the bill of materials by omitting an external EEPROM or flash part.
Fortunately, such NICs also include the ability to autoload the MAC
address in hardware.

If the autoloaded MAC address is valid, use it in preference to
attempting to read a MAC address from a potentially non-existent
EEPROM.

Originally-fixed-by: Donald Post <Donald.Post@utas.utc.com>
Signed-off-by: Michael Brown <mcb30@ipxe.org>