[tls] Send empty Certificate record if we have no private key
We currently treat a request for a client certificate as fatal if we
do not have both a private key and a corresponding certificate.  This
prevents the use of optional client certificates by HTTP servers.

Fix by sending an empty Certificate record if the server requests a
client certificate but we have no private key available.  This allows
the server to decide whether or not to proceed without a client
certificate.

If the server requests a client certificate and we have a private key
but no matching certificate available, then continue to treat this as
a fatal error since it almost certainly indicates a user error in the
build process (or other misconfiguration).

Signed-off-by: Michael Brown <mcb30@ipxe.org>