[intelxl] Allow for supporting multiple API versions
The XL710 and E810 have very similar firmware APIs, and can viably be
considered as different versions of the same API.  Both are nominally
version 1, but the layout of the response to the initial "get API
version" call can be used to disambiguate between them.

Add a skeleton abstraction of an API version.

Signed-off-by: Michael Brown <mcb30@ipxe.org>