[intelxl] Add support for E810 version of Get Switch Configuration
The E810 has an updated layout for the data buffer used by the Get
Switch Configuration command.  Handle this difference via the API
version abstraction.

The XL710 places the 16-bit "next element" continuation field in bytes
16-17 of the command parameters, with bytes 18-19 as "reserved, must
be zero".  The E810 places the 16-bit "next element" continuation
field in bytes 18-19 of the command parameters, with bytes 16-17 as
"reserved, must be zero".  We can handle this cleanly by treating the
field as an opaque 32-bit value for both API versions.

Signed-off-by: Michael Brown <mcb30@ipxe.org>