[intelxl] Add support for E810 version of Get Link Status
The E810 has an updated layout for the command parameters used by the
Get Link Status command, and returns the result via the data buffer.
Handle this difference via the API version abstraction.

The XL710 places the "link status notification" in byte 16 of the
command parameters, with byte 18 defined as "value is ignored".  The
E810 places the "link status notification" in byte 18 of the command
parameters, with byte 16 defined as "value is ignored".  We can handle
this cleanly by populating both fields.

Signed-off-by: Michael Brown <mcb30@ipxe.org>