[intelxl] Use admin queue to get port MAC address and maximum frame size
Remove knowledge of the PRTPM_SA[HL] registers, and instead use the
admin queue to retrieve the MAC address.

The maximum frame size is already retrieved via the admin queue as
part of the link state information.  Extend the intelxl_admin_link()
function to allow this to be used to populate the network device's
maximum packet length.

Signed-off-by: Michael Brown <mcb30@ipxe.org>