[ena] Specify the unused completion queue MSI-X vector as 0xffffffff
Some versions of the ENA firmware will complain if the completion
queue's MSI-X vector field is left empty, even though the queue
configuration specifies that interrupts are not used.

Work around these firmware versions by passing in what appears to be
the magic "no MSI-X vector" value in this field.

Signed-off-by: Michael Brown <mcb30@ipxe.org>