[crypto] Support in-place decryption for GCM ciphers
The hash calculation is performed incorrectly when decrypting in
place, since the ciphertext will have been overwritten with the
plaintext before being used to update the hash value.

Restructure the code to allow for in-place encryption and decryption.
Choose to optimise for the decryption case, since we are likely to
decrypt much more data than we encrypt.

Signed-off-by: Michael Brown <mcb30@ipxe.org>