[efi] Add efi_path_family() to parse an address family from a device path
EFI requires an explicit choice to be made between the use of IPv4 and
IPv6 address families even in contexts where making such a choice
violates basic causality.  For example: initialisation of an
EFI_HTTP_PROTOCOL instance requires the caller to make a premature
up-front choice between using IPv4 or IPv6, before performing the DNS
resolution to find out to which address family (or families) the HTTP
server name resolves.

Provide efi_path_family() to scan a device path to check for the
presence of an IPv4 or IPv6 path element.

Signed-off-by: Michael Brown <mcb30@ipxe.org>