[crypto] Use Montgomery reduction for modular exponentiation
Speed up modular exponentiation by using Montgomery reduction rather
than direct modular reduction.

This implementation imposes a limitation that the modulus must be an
odd number.  This is not a limitation in practice, since in all
current use cases the modulus is a large prime (and is therefore odd).

If modular exponentation with an even modulus is required in future,
this could be implemented by using Garner's algorithm as described in
"Montgomery reduction with even modulus" (Koç, 1994):

  - right-shift the modulus until it is odd, calculating m = q * 2^j
    where q is odd

  - calculate x1 = base ^ exp mod q using Montgomery reduction

  - calculate x2 = base ^ exp mod 2^j (by truncating at each step)

  - calculate q^-1 mod 2^j via extended bigint_mod_invert()

  - calculate y = (x2 - x1) * q^-1 mod 2^j

  - calculate result as x = x1 + q * y

Signed-off-by: Michael Brown <mcb30@ipxe.org>