[crypto] Support modular exponentiation only with an odd modulus
Performing modular exponentiation is substantially faster when using
Montgomery reduction rather than direct modular reduction.  Montgomery
reduction in base 2^n requires the modulus to be coprime to 2^n, which
limits us to requiring that the modulus is an odd number.

This is not a limitation in practice, since in all current use cases
the modulus is a large prime (and is therefore odd).

Adjust all modular exponentiation test cases to have an odd modulus,
in preparation for using Montgomery reduction.

Signed-off-by: Michael Brown <mcb30@ipxe.org>