[i2c] Remove device address length field
We choose to represent a 10-bit I2C device address as the 15-bit value
actually sent on the wire (i.e. as binary 1111xxxxxxxxxx).  There is
therefore no overlap between the 7-bit and 10-bit namespaces, and we
can therefore deduce the address length from the address alone.

Remove the separate device address length field to simplify the code.

Signed-off-by: Michael Brown <mcb30@ipxe.org>