[gve] Add support for raw DMA addressing
Raw DMA addressing allows the transmit and receive descriptors to
provide the DMA address of the data buffer directly, without requiring
the use of a pre-registered queue page list.  It is modelled in the
device as a magic "raw DMA" queue page list (with QPL ID 0xffffffff)
covering the whole of the DMA address space.

When using raw DMA addressing, the transmit and receive datapaths
could use the normal pattern of mapping I/O buffers directly, and
avoid copying packet data into and out of the fixed queue page list
ring buffer.  However, since we must retain support for queue page
list addressing (which requires this additional copying), we choose to
minimise code size by continuing to use the fixed ring buffer even
when using raw DMA addressing.

Add support for raw DMA addressing and for selecting the mode of
operation based on options advertised as supported by the device.

Note that there are no known devices that actually advertise support
for the GQI-RDA mode of operation, and so this remains untested in
practice.  It is, however, a prerequisite for supporting the DQO-RDA
mode of operation (i.e. out-of-order descriptor queues with raw DMA
addressing) as required by some newer machine types.

Signed-off-by: Michael Brown <mcb30@ipxe.org>