[ena] Check for a broken prefetchable memory BAR
On some AMD-based systems (observed on a c7a.medium instance in
eu-west-2), the prefetchable memory BAR appears to be non-functional
even though all intermediate PCI bridges are correctly configured (and
the BAR value has been assigned by iPXE based on the immediately
containing bridge).  The symptom is that ringing the transmit doorbell
will immediately lock up the hardware (as happens if the doorbell is
rung without having actually written a valid transmit descriptor).

On systems exhibiting this problem, reads from the memory BAR return
an all-ones bit pattern typical of a PCI master abort or other
nonexistent device read.  There is no direct way to test whether or
not writes are reaching the device, but we can exploit the read
behaviour to detect the problem and work around it by disabling the
use of LLQs.

Signed-off-by: Michael Brown <mcb30@ipxe.org>