[efi] Allow for creating devices with no EFI parent device
When running on systems where the UEFI firmware fails to enumerate
some of the underlying hardware devices (such as an AWS m8g.medium
instance), we cannot comply with the UEFI device model by adding our
SNP device as a child of the hardware device and appending to the
parent hardware device path.

Work around these systems by allowing for the creation of SNP devices
with no parent device.

Signed-off-by: Michael Brown <mcb30@ipxe.org>